Forward progress on potentials.
